isolate geohash layer in separate module
